+++ sapl/base/urls.py
@@ -3,31 +3,22 @@ import os
 from django.conf.urls import include, url
 from django.contrib.auth import views
 from django.contrib.auth.decorators import permission_required
-
 from django.views.generic.base import RedirectView, TemplateView
 
 from sapl.base.views import (AutorCrud, ConfirmarEmailView, TipoAutorCrud, get_estatistica,
                              RecuperarSenhaEmailView, RecuperarSenhaFinalizadoView,
-                             RecuperarSenhaConfirmaView, RecuperarSenhaCompletoView, RelatorioMateriaAnoAssuntoView,
-                             IndexView, UserCrud)
+                             RecuperarSenhaConfirmaView, RecuperarSenhaCompletoView, IndexView, UserCrud)
 from sapl.settings import MEDIA_URL, LOGOUT_REDIRECT_URL
-
 from .apps import AppConfig
-from .forms import LoginForm
 from .views import (LoginSapl, AlterarSenha, AppConfigCrud, CasaLegislativaCrud,
-                    HelpTopicView, LogotipoView, RelatorioAtasView, PesquisarAuditLogView,
-                    RelatorioAudienciaView, RelatorioDataFimPrazoTramitacaoView, RelatorioHistoricoTramitacaoView,
-                    RelatorioMateriasPorAnoAutorTipoView, RelatorioMateriasPorAutorView,
-                    RelatorioMateriasTramitacaoView, RelatorioPresencaSessaoView, RelatorioReuniaoView, SaplSearchView,
-                    RelatorioNormasPublicadasMesView, RelatorioNormasVigenciaView,
-                    EstatisticasAcessoNormas, RelatoriosListView, ListarInconsistenciasView,
+                    HelpTopicView, LogotipoView, PesquisarAuditLogView,
+                    SaplSearchView,
+                    ListarInconsistenciasView,
                     ListarProtocolosDuplicadosView, ListarProtocolosComMateriasView, ListarMatProtocoloInexistenteView,
                     ListarParlamentaresDuplicadosView, ListarFiliacoesSemDataFiliacaoView,
                     ListarMandatoSemDataInicioView, ListarParlMandatosIntersecaoView, ListarParlFiliacoesIntersecaoView,
                     ListarAutoresDuplicadosView, ListarBancadaComissaoAutorExternoView, ListarLegislaturaInfindavelView,
-                    ListarAnexadasCiclicasView, ListarAnexadosCiclicosView, pesquisa_textual,
-                    RelatorioHistoricoTramitacaoAdmView, RelatorioDocumentosAcessoriosView, RelatorioNormasPorAutorView)
-
+                    ListarAnexadasCiclicasView, ListarAnexadosCiclicosView, pesquisa_textual)
 
 app_name = AppConfig.name
 
@@ -67,53 +58,6 @@ urlpatterns = [
     url(r'^sistema/casa-legislativa/', include(CasaLegislativaCrud.get_urls()),
         name="casa_legislativa"),
     url(r'^sistema/app-config/', include(AppConfigCrud.get_urls())),
-
-    # TODO mover estas telas para a app 'relatorios'
-    url(r'^sistema/relatorios/$',
-        RelatoriosListView.as_view(), name='relatorios_list'),
-    url(r'^sistema/relatorios/materia-por-autor$',
-        RelatorioMateriasPorAutorView.as_view(), name='materia_por_autor'),
-    url(r'^sistema/relatorios/relatorio-por-mes$',
-        RelatorioNormasPublicadasMesView.as_view(), name='normas_por_mes'),
-    url(r'^sistema/relatorios/relatorio-por-vigencia$',
-        RelatorioNormasVigenciaView.as_view(), name='normas_por_vigencia'),
-    url(r'^sistema/relatorios/estatisticas-acesso$',
-        EstatisticasAcessoNormas.as_view(), name='estatisticas_acesso'),
-    url(r'^sistema/relatorios/materia-por-ano-autor-tipo$',
-        RelatorioMateriasPorAnoAutorTipoView.as_view(),
-        name='materia_por_ano_autor_tipo'),
-    url(r'^sistema/relatorios/materia-por-tramitacao$',
-        RelatorioMateriasTramitacaoView.as_view(),
-        name='materia_por_tramitacao'),
-    url(r'^sistema/relatorios/materia-por-assunto$',
-        RelatorioMateriaAnoAssuntoView.as_view(),
-        name='materia_por_ano_assunto'),
-    url(r'^sistema/relatorios/historico-tramitacoes$',
-        RelatorioHistoricoTramitacaoView.as_view(),
-        name='historico_tramitacoes'),
-    url(r'^sistema/relatorios/data-fim-prazo-tramitacoes$',
-        RelatorioDataFimPrazoTramitacaoView.as_view(),
-        name='data_fim_prazo_tramitacoes'),
-    url(r'^sistema/relatorios/presenca$',
-        RelatorioPresencaSessaoView.as_view(),
-        name='presenca_sessao'),
-    url(r'^sistema/relatorios/atas$',
-        RelatorioAtasView.as_view(),
-        name='atas'),
-    url(r'^sistema/relatorios/reuniao$',
-        RelatorioReuniaoView.as_view(),
-        name='reuniao'),
-    url(r'^sistema/relatorios/audiencia$',
-        RelatorioAudienciaView.as_view(),
-        name='audiencia'),
-    url(r'^sistema/relatorios/historico-tramitacoesadm$',
-        RelatorioHistoricoTramitacaoAdmView.as_view(),
-        name='historico_tramitacoes_adm'),
-    url(r'^sistema/relatorios/documentos_acessorios$',
-        RelatorioDocumentosAcessoriosView.as_view(),
-        name='relatorio_documentos_acessorios'),
-    url(r'^sistema/relatorios/normas-por-autor$',
-        RelatorioNormasPorAutorView.as_view(), name='normas_por_autor'),
 
     url(r'^email/validate/(?P<uidb64>[0-9A-Za-z_\-]+)/'
         '(?P<token>[0-9A-Za-z]{1,13}-[0-9A-Za-z]{1,20})$',
