+++ sapl/relatorios/urls.py
@@ -4,10 +4,15 @@ from .apps import AppConfig
 from .views import (relatorio_capa_processo,
                     relatorio_documento_administrativo, relatorio_espelho,
                     relatorio_etiqueta_protocolo, relatorio_materia,
-                    relatorio_ordem_dia, relatorio_pauta_sessao,
-                    relatorio_protocolo, relatorio_sessao_plenaria,
+                    relatorio_ordem_dia, relatorio_protocolo, relatorio_sessao_plenaria,
                     resumo_ata_pdf, relatorio_sessao_plenaria_pdf, etiqueta_materia_legislativa,
-                    relatorio_materia_tramitacao)
+                    relatorio_materia_tramitacao, RelatoriosListView, RelatorioMateriasPorAutorView,
+                    RelatorioNormasPublicadasMesView, RelatorioNormasVigenciaView, RelatorioMateriasPorAnoAutorTipoView,
+                    RelatorioMateriasTramitacaoView, RelatorioMateriaAnoAssuntoView, RelatorioHistoricoTramitacaoView,
+                    RelatorioDataFimPrazoTramitacaoView, RelatorioPresencaSessaoView, RelatorioAtasView,
+                    RelatorioReuniaoView, RelatorioAudienciaView, RelatorioHistoricoTramitacaoAdmView,
+                    RelatorioDocumentosAcessoriosView, RelatorioNormasPorAutorView)
+from ..base.views import EstatisticasAcessoNormas
 
 app_name = AppConfig.name
 
@@ -45,4 +50,51 @@ urlpatterns = [
 
     url(r'^relatorios/(?P<pk>\d+)/materia-tramitacao$',
         relatorio_materia_tramitacao, name='relatorio_materia_tramitacao'),
+
+    # TODO mover estas telas para a app 'relatorios'
+    url(r'^sistema/relatorios/$',
+        RelatoriosListView.as_view(), name='relatorios_list'),
+    url(r'^sistema/relatorios/materia-por-autor$',
+        RelatorioMateriasPorAutorView.as_view(), name='materia_por_autor'),
+    url(r'^sistema/relatorios/relatorio-por-mes$',
+        RelatorioNormasPublicadasMesView.as_view(), name='normas_por_mes'),
+    url(r'^sistema/relatorios/relatorio-por-vigencia$',
+        RelatorioNormasVigenciaView.as_view(), name='normas_por_vigencia'),
+    url(r'^sistema/relatorios/estatisticas-acesso$',
+        EstatisticasAcessoNormas.as_view(), name='estatisticas_acesso'),
+    url(r'^sistema/relatorios/materia-por-ano-autor-tipo$',
+        RelatorioMateriasPorAnoAutorTipoView.as_view(),
+        name='materia_por_ano_autor_tipo'),
+    url(r'^sistema/relatorios/materia-por-tramitacao$',
+        RelatorioMateriasTramitacaoView.as_view(),
+        name='materia_por_tramitacao'),
+    url(r'^sistema/relatorios/materia-por-assunto$',
+        RelatorioMateriaAnoAssuntoView.as_view(),
+        name='materia_por_ano_assunto'),
+    url(r'^sistema/relatorios/historico-tramitacoes$',
+        RelatorioHistoricoTramitacaoView.as_view(),
+        name='historico_tramitacoes'),
+    url(r'^sistema/relatorios/data-fim-prazo-tramitacoes$',
+        RelatorioDataFimPrazoTramitacaoView.as_view(),
+        name='data_fim_prazo_tramitacoes'),
+    url(r'^sistema/relatorios/presenca$',
+        RelatorioPresencaSessaoView.as_view(),
+        name='presenca_sessao'),
+    url(r'^sistema/relatorios/atas$',
+        RelatorioAtasView.as_view(),
+        name='atas'),
+    url(r'^sistema/relatorios/reuniao$',
+        RelatorioReuniaoView.as_view(),
+        name='reuniao'),
+    url(r'^sistema/relatorios/audiencia$',
+        RelatorioAudienciaView.as_view(),
+        name='audiencia'),
+    url(r'^sistema/relatorios/historico-tramitacoesadm$',
+        RelatorioHistoricoTramitacaoAdmView.as_view(),
+        name='historico_tramitacoes_adm'),
+    url(r'^sistema/relatorios/documentos_acessorios$',
+        RelatorioDocumentosAcessoriosView.as_view(),
+        name='relatorio_documentos_acessorios'),
+    url(r'^sistema/relatorios/normas-por-autor$',
+        RelatorioNormasPorAutorView.as_view(), name='normas_por_autor'),
 ]
