+++ sapl/templates/navbar.yaml
@@ -57,7 +57,7 @@
       url: sapl.materia:proposicao_list
       check_permission: materia.add_proposicao
     - title: {% trans 'Relatórios' %}
-      url: sapl.base:relatorios_list
+      url: sapl.relatorios:relatorios_list
     - title: {% trans 'Sessões Plenárias' %}
       url: sapl.sessao:pesquisar_sessao
     - title: {% trans 'Tramitação em Lote' %}
