+++ sapl/templates/relatorios/RelatorioAtas_filter.html
@@ -10,7 +10,7 @@
 
   	{% if filter_url %}
 		<div class="actions btn-group float-right" role="group">
-			<a href="{% url 'sapl.base:atas' %}" class="btn btn-outline-primary">{% trans 'Fazer nova pesquisa' %}</a>
+			<a href="{% url 'sapl.relatorios:atas' %}" class="btn btn-outline-primary">{% trans 'Fazer nova pesquisa' %}</a>
 		</div>
 		<br/><br/><br/>
 		<b>PERÍODO: {{ periodo }}<br /></b><br /><br/>
