+++ sapl/templates/relatorios/RelatorioDocumentosAcessorios_filter.html
@@ -7,7 +7,7 @@
         {% crispy filter.form %}
     {% else %}
         <div class="actions btn-group float-right" role="group">
-            <a href="{% url 'sapl.base:relatorio_documentos_acessorios' %}" class="btn btn-outline-primary">{% trans 'Fazer uma nova pesquisa' %}</a>
+            <a href="{% url 'sapl.relatorios:relatorio_documentos_acessorios' %}" class="btn btn-outline-primary">{% trans 'Fazer uma nova pesquisa' %}</a>
         </div>
         <br /><br /><br /><br />
         <b>PARÂMETROS DE PESQUISA<br /></b>
