+++ sapl/templates/relatorios/RelatorioHistoricoTramitacaoAdm_filter.html
@@ -8,7 +8,7 @@
     {% crispy filter.form %}
   {% else %}
     <div class="actions btn-group float-right" role="group">
-      <a href="{% url 'sapl.base:historico_tramitacoes_adm' %}" class="btn btn-outline-primary">{% trans 'Fazer nova pesquisa' %}</a>
+      <a href="{% url 'sapl.relatorios:historico_tramitacoes_adm' %}" class="btn btn-outline-primary">{% trans 'Fazer nova pesquisa' %}</a>
     </div>
     <br /><br /><br /><br />
     <b>PARÂMETROS DE PESQUISA:<br /></b>
