+++ sapl/templates/relatorios/RelatorioNormasVigencia_filter.html
@@ -9,7 +9,7 @@
 
   {% if show_results %}
     <div class="actions btn-group float-right" role="group">
-      <a href="{% url 'sapl.base:normas_por_vigencia' %}" class="btn btn-outline-primary">{% trans 'Fazer nova pesquisa' %}</a>
+      <a href="{% url 'sapl.relatorios:normas_por_vigencia' %}" class="btn btn-outline-primary">{% trans 'Fazer nova pesquisa' %}</a>
     </div>
     <br /><br /><br /><br />
     <b>PARÂMETROS DE PESQUISA:<br /></b>
