+++ sapl/templates/relatorios/RelatorioPresencaSessao_filter.html
@@ -16,7 +16,7 @@
     </style>
 
     <div class="actions btn-group float-right" role="group">
-      <a href="{% url 'sapl.base:presenca_sessao' %}" class="btn btn-outline-primary">{% trans 'Fazer nova pesquisa' %}</a>
+      <a href="{% url 'sapl.relatorios:presenca_sessao' %}" class="btn btn-outline-primary">{% trans 'Fazer nova pesquisa' %}</a>
     </div>
     <br /><br /><br /><br />
     <b>PERÍODO: {{periodo}}</b><br />
