+++ sapl/templates/relatorios/RelatorioReuniao_filter.html
@@ -9,7 +9,7 @@
 
   {% if show_results %}
     <div class="actions btn-group float-right" role="group">
-      <a href="{% url 'sapl.base:reuniao' %}" class="btn btn-outline-primary">{% trans 'Fazer nova pesquisa' %}</a>
+      <a href="{% url 'sapl.relatorios:reuniao' %}" class="btn btn-outline-primary">{% trans 'Fazer nova pesquisa' %}</a>
     </div>
     <br /><br /><br /><br />
     {% if object_list|length > 0 %}
